Work around broken UPC-A barcodes
Some old printed barcodes are only 12 characters long, but begin with
a zero. As a result, a checksum is calculated, making the barcode
invalid e.g.

012345678912x is scanned in as 12345678912x, where x is the checksum as
our scanner discards the first digit if 0

As a result, those old barcodes have become 0012345678912 in the database
in order to stop the checksum being calculated (they are now 13 chars)

In order to work around this, we check for the '00' + scanned_barcode as
a second criteria